Add in db configuration / management
Add in the basic database configuration export management. This has the
shell class for PostgreSQL but doesn't actually do anything. Something
similar to how puppetlabs/puppetdb does things should be implemented for
the PostgreSQL configuration. It would be great to do the same for MySQL
configuration but some of the needed bits aren't in the puppetlabs/mysql
module :(

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>