Silence unneeded wget::fetch notices
The wget::fetch was issuing a notice on every puppet run. This only
happens when a cache_dir is set. Since the plugins are downloading a
single file we switch to using the --timestamping feature of wget to
ensure it only downloads a new version if it has to and download
directly in place.

This kills off the unneeded notice that was being emitted.

Additionally, because we were using a cache dir previously and it was
being created (if needed) we now purge it from the system if it exists.
This purge will be removed at or after v1.0.0 releases.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>