Allow Java options tuning on systemd
We need to be able to tune the JAVA_OPTIONS if we're on a systemd based
system (F21+ or EL7+). This modifies the /etc/defaults/gerritcodereview
environment file on systemd based systems to add in a JAVA_OPTIONS
environment line that tunes the 2 options that the SysV init script can
tune for JAVA_OPTIONS

We do it in this file as it's sourced by the service unit file. We're
doing this instead of rewriting the unit file since doing so requires a
reload of systemd if it's changed and adding that to this module is
outside of the scope.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>